# PHASE-2/backend/src/schemas/task.py
from datetime import datetime
from uuid import UUID

from pydantic import BaseModel, Field


class TaskCreate(BaseModel):
    title: str = Field(min_length=1, max_length=500)
    description: str | None = None


class TaskUpdate(BaseModel):
    title: str = Field(min_length=1, max_length=500)
    description: str | None = None


class TaskRead(BaseModel):
    id: UUID
    user_id: str
    title: str
    description: str | None
    is_completed: bool
    completed_at: datetime | None
    created_at: datetime
    updated_at: datetime

    model_config = {"from_attributes": True}