# PHASE-2/specs/001-todo-backend-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Todo Full-Stack Web Application API
  version: 1.0.0
  description: RESTful API for managing user tasks with JWT authentication

servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.example.com
    description: Production

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          example: "Buy groceries"
        description:
          type: string
          nullable: true
          example: "Milk, eggs, bread"

    TaskUpdate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        description:
          type: string
          nullable: true

    TaskRead:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        is_completed:
          type: boolean
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string
          nullable: true

    SuccessListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/TaskRead'
        message:
          type: string
          nullable: true

    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        message:
          type: string

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
          nullable: true
        code:
          type: string

    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Task deleted successfully"

security:
  - BearerAuth: []

paths:
  /api/{user_id}/tasks:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: List all tasks for user
      operationId: listTasks
      tags: [Tasks]
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessListResponse'
              example:
                success: true
                data:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    user_id: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                    title: "Buy groceries"
                    description: "Milk, eggs, bread"
                    is_completed: false
                    completed_at: null
                    created_at: "2026-02-09T10:00:00Z"
                    updated_at: "2026-02-09T10:00:00Z"
                message: null
        '401':
          description: Unauthorized - missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user_id mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create a new task
      operationId: createTask
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TaskRead'
              example:
                success: true
                data:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  user_id: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                  title: "Buy groceries"
                  description: "Milk, eggs, bread"
                  is_completed: false
                  completed_at: null
                  created_at: "2026-02-09T10:00:00Z"
                  updated_at: "2026-02-09T10:00:00Z"
                message: "Task created successfully"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Validation failed"
                details:
                  - field: "title"
                    message: "field required"
                code: "VALIDATION_ERROR"

  /api/{user_id}/tasks/{task_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: Get a single task
      operationId: getTask
      tags: [Tasks]
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TaskRead'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Task not found"
                details: null
                code: "NOT_FOUND"

    put:
      summary: Update a task (full replacement)
      operationId: updateTask
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TaskRead'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete a task
      operationId: deleteTask
      tags: [Tasks]
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
              example:
                success: true
                message: "Task deleted successfully"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/{user_id}/tasks/{task_id}/complete:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    patch:
      summary: Mark a task as complete
      operationId: completeTask
      tags: [Tasks]
      responses:
        '200':
          description: Task marked as complete
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TaskRead'
              example:
                success: true
                data:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  user_id: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                  title: "Buy groceries"
                  description: "Milk, eggs, bread"
                  is_completed: true
                  completed_at: "2026-02-09T12:00:00Z"
                  created_at: "2026-02-09T10:00:00Z"
                  updated_at: "2026-02-09T12:00:00Z"
                message: "Task marked as complete"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'